+++ Sources/orrery/main.swift
@@ -1,2 +1,4 @@
 import OrreryCore
+
+LegacyOrbitalMigration.runIfNeeded()
 OrreryCommand.main()